requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
